plugins/gemini/pyproject.toml
[project]
name = "flyteplugins-gemini"
dynamic = ["version"]
description = "Google Gemini plugin for flyte"
readme = "README.md"
authors = [{ name = "Andre Ahlert", email = "andre@aex.partners" }]
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
    "flyte",
]

[build-system]
requires = ["setuptools", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
]

[tool.setuptools]
include-package-data = true
license-files = ["licenses/*.txt", "LICENSE"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["flyteplugins*"]

[tool.setuptools_scm]
root = "../../"

[tool.pytest.ini_options]
norecursedirs = []
log_cli = true
log_cli_level = 20
markers = []
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
branch = true

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "PLW",
    "YTT",
    "ASYNC",
    "C4",
    "T10",
    "EXE",
    "ISC",
    "LOG",
    "PIE",
    "Q",
    "RSE",
    "FLY",
    "PGH",
    "PLC",
    "PLE",
    "PLW",
    "FURB",
    "RUF",
]
ignore = ["PGH003", "PLC0415"]

[tool.ruff.lint.per-file-ignores]
"examples/*" = ["E402"]

[tool.uv.sources]
flyte = { path = "../../", editable = true }
